Added functionality to show division result while 'Action' button is pressed
Modified the code to display the result of dividing the current input by 9 when the 'Action' button is pressed and to restore the original number when the button is released. This allows the user to see the division result while holding down the button and revert to the original number upon release.

Fixes #<issue_number> (if applicable)